feat: show logs and command acks in terminal
